python-function/requirements.txt
azure-functions==1.20.0
azure-cosmos==4.5.1
PyJWT==2.9.0
